extract `reduce_simple_command` logic
